+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/deploy-tooling
 
+## 0.16.4
+
+### Patch Changes
+
+-   Updated dependencies [61d4060]
+    -   @sap-ux/axios-extension@1.21.2
+    -   @sap-ux/system-access@0.6.2
+
 ## 0.16.3
 
 ### Patch Changes